+++ stroma/models/aurum/gold__person.sql
@@ -29,20 +29,20 @@ MODEL (
 
 /* This is the patient table. */
 SELECT
-  p.person_id::INT,
-  p.gender_concept_id::INT,
-  p.year_of_birth::INT, /* This is the year the patient was born */
-  p.month_of_birth::INT,
-  p.race_concept_id::INT,
-  p.ethnicity_concept_id::INT,
-  p.location_id::INT,
-  p.provider_id::INT,
-  p.care_site_id::INT,
+  p.person_id::BIGINT,
+  p.gender_concept_id::BIGINT,
+  p.year_of_birth::BIGINT, /* This is the year the patient was born */
+  p.month_of_birth::BIGINT,
+  p.race_concept_id::BIGINT,
+  p.ethnicity_concept_id::BIGINT,
+  p.location_id::BIGINT,
+  p.provider_id::BIGINT,
+  p.care_site_id::BIGINT,
   p.person_source_value::TEXT,
   p.gender_source_value::TEXT,
-  p.gender_source_concept_id::INT,
+  p.gender_source_concept_id::BIGINT,
   p.race_source_value::TEXT,
-  p.race_source_concept_id::INT,
+  p.race_source_concept_id::BIGINT,
   p.ethnicity_source_value::TEXT,
-  p.ethnicity_source_concept_id::INT
+  p.ethnicity_source_concept_id::BIGINT
 FROM stg_gold.stg__person AS p
